Update additional lockfiles after bumping importmap-rails version
Without this, running `bundle install` with `BUNDLE_GEMFILE` set to one
of these gemfiles will result in an error like:
```
You are trying to install in deployment mode after changing your
Gemfile. Run `bundle install` elsewhere and add the updated
gemfiles/rails_7_propshaft.gemfile.lock to version control.
```
as seen in the CI run for the commit that bumped the version:
https://github.com/rails/importmap-rails/runs/4297724678?check_suite_focus=true#step:3:30