Add package information in vendored package file
Now that the packages are downloaded by default, apart from the output of the `pin` command we lose the track of the version and from where the package was downloaded. To track the version the solution was use a comment in bin/importmap next to the package version but nothing for the package origin so the proposal is add a comment in the first line of the downloaded package displaying the package name, version and from where it was downloaded.