Add breaking test
See https://github.com/rails/importmap-rails/issues/245 for context.

The TL;DR is the response sometimes returns a String instead of Hash object. This should be handled properly.

The way to handle this is up for discussion. To start addressing the issue the String response will just be surfaced as `latest_version`.