Make integrity calculation opt-in
This commits introduces a new `enable_integrity!` method to the importmap
configuration, allowing users to opt-in to automatic integrity hash
calculation. This change ensures that integrity hashes
are not calculated by default, giving developers control over when to
enable this feature.

This is already opt-in for Propshaft users, but now it is also opt-in for
Sprockets users.